Differentiate hiker and trail_angel posts

+++ PostService.py
@@ -1,15 +1,27 @@
 from Post import Post
 from HikerService import HikerService
+from TrailAngelService import TrailAngelService
+from UserType import UserType
 class PostService:
 
     def __init__(self, db) -> None:
         self.db = db
 
     def create_post(self, post):
-        hiker = HikerService(self.db).get_hiker(post.hiker_id)
 
-        if not hiker:
-            raise Exception("Problem validating hiker creating post")
+        user = None
+        user_type = UserType(post.user_type)
+
+        if(user_type == UserType.HIKER):
+            user = HikerService(self.db).get_hiker(post.user_id)
+        elif(user_type == UserType.TRAILANGEL):
+            user = TrailAngelService(self.db).get_trail_angel(post.user_id)
+        else:
+            raise Exception(f"User Type [{post.user_type}] Undefined")
+
+        if not user:
+            raise Exception("Problem validating user creating post")
+
         row = self.db.save_post(post)
         return self.__post_from_row(row)
 
@@ -25,6 +37,10 @@ class PostService:
         return post
 
     def delete_post(self, post_id):
+        existing_post = self.db.get_post(post_id)
+        if not existing_post:
+            print(f"Attempted to delete nonexistent post [{post_id}]")
+            return
         self.db.delete_post(post_id)
 
     def __posts_from_rows(self, rows):
